fix: preparation for adding profile fragment
